pyproject.toml
[tool.poetry]
name = "validation-service"
version = "1.0.0"
description = ""
authors = ["Matvei Stepin [email: matveystepin07@gmail.com]"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
